Create handlers for vote create/destroy ajax response
For some reason, ajax:success is not being triggered.

The ajax:complete action is triggered, and if we inspect its status we
see:

    status
    -> 'parseerror'

This is likely because jQuery expects the response to be javascript, and
it attempts to execute it. We're sending html, so javascript is not able
to parse it, hence the error.